continuous encryption and decryption process [bin] added
