+++ Cargo.toml
@@ -1,8 +1,9 @@
 [package]
 name = "poems"
 version = "0.1.0"
-edition = "2021"
+edition = "2018"
 
+# crate-type = ["cdylib"]
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
@@ -49,10 +50,26 @@ wasm-bindgen = "0.2.87"
 webbrowser = "0.8.10"
 
 [lib]
-crate-type = ["cdylib"]
+name = "poems"
+path = "src/lib.rs"
+
+
 
 [profile.dev.package.num-bigint-dig]
 opt-level = 3
 
 [dev-dependencies]
 rusty-hook = "^0.11.2"
+
+
+[[bin]]
+name = "decrypt"
+path = "src/bin/decrypt.rs"
+
+[[bin]]
+name = "registration"
+path = "src/bin/registration.rs"
+
+
+
+
