+++ record.json
@@ -1,0 +1,82 @@
+{
+  "keys": {
+    "e": "991b1e1eea36b0e54401beff982cc6516bb72222ec0d25a2971e8e0dc9a6ddc7",
+    "m": "october suggest marble stamp help increase avoid daughter zebra scrap cover people rocket much merry like certain believe sibling bright symptom square jacket ugly",
+    "d": "215fe4fb1f89e4d0913f7c77a47c42e66ceb01fe55d0a463d8cd8546d130052b5e02ba8a8eea9d096f26cb27282271528eb4e8030b63ab7539f3601a16f655c7",
+    "p": "MIIBIjANBgkqhkiG9w0BAQEFAAOCAQ8AMIIBCgKCAQEAs8BOxrevs6SQ6rI6SV0lanNcghpRoNO6ixPzPUUVR+Fk5IXG93zNABYEj1tYwTo/HADbZ3WEje/Ri8NelUicN3PaVHxM49RM5zx8FEy4pNdzs9EaMzei5xfDa7GCvvdC2A0jT0m6yAYZHCnFkvSITgrhLO/rRet8GoJ/6aT9wVQS2loTcWEOXbSudXEGDe20kQWisVVUNEc5Y236vHxX7Gy0Ja7DwyFxgY+9fjYEXEsvXLxk3JzJLEOLiuIaoMzudSZL4R08mEodX1jfXd8zRswTMncnTZKCSa3LvlKxhDNxsXZCKELoBiLPLCRIxdcPhxq9gBZhHZtSapETQY7uTwIDAQAB",
+    "pk": "MIIEvQIBADANBgkqhkiG9w0BAQEFAASCBKcwggSjAgEAAoIBAQCzwE7Gt6+zpJDqsjpJXSVqc1yCGlGg07qLE/M9RRVH4WTkhcb3fM0AFgSPW1jBOj8cANtndYSN79GLw16VSJw3c9pUfEzj1EznPHwUTLik13Oz0RozN6LnF8NrsYK+90LYDSNPSbrIBhkcKcWS9IhOCuEs7+tF63wagn/ppP3BVBLaWhNxYQ5dtK51cQYN7bSRBaKxVVQ0Rzljbfq8fFfsbLQlrsPDIXGBj71+NgRcSy9cvGTcnMksQ4uK4hqgzO51JkvhHTyYSh1fWN9d3zNGzBMydydNkoJJrcu+UrGEM3GxdkIoQugGIs8sJEjF1w+HGr2AFmEdm1JqkRNBju5PAgMBAAECggEAdl2ogkrrNIylsSb0IuNnhBW7USErp5BFQtg8UJhxTwgpfCJrxwUzqV7WWeB5LaTBchtOX26Pfzfw8OwSGJPb6tgraX8A0vY2Pf8iZeuvmz0VP8/uRNPaPfr0ALK/uijwp3zfR5LxoE4Dsvv+0qqcJ2gap5DFiSFsc7iDqsOvJeKwcs/A5+8+QMC9Su1h/gC64iCwdaMebY6ECUMS79g/EH0BRGN0/JrzbZx2hxNBBoFrLyV893dOt9ZfYV0L1odyacDBUAFduP78t8ZrrLcqT/MAdDe6Y93sZlJXs3ouYZkpfLkhnFiYqs8e0uLn6sP6M89mqpoH6GsyFNCvjQvz4QKBgQDQPDwle+I/WRxkrsG1s6tXa2dcKtIJlFXSM09BOT1StVBf0w1eVrtKLCQZpUPanngGtbzZhn6NRpIgEncY1k8nWZg+HJ0jvgA/f55KpTg8BU+LqXQzuskSk9OzTzmZlBhTIhUktQD+xRMUVWJi5iFffHGiT0j644ZZq9OnvnqkvQKBgQDc+3YOd87UbYyaGNCZP4yaZTYkaikZ+ofH53JjUNRMg18i+/jTQmas6xKizmBg3+hEPpdzDp/YrB9tRnfiVJFrOehkLhLR6oOiRuEF+sPTH551OM9sO2UvPYL6eoYQiT+SIJokm3aOqidILIX9/sUnCcXrvypZqQnsYf63hp4d+wKBgFacpzvEEm/c5bVvBr1H9P2Qe9ZXLTO00hlzJATyBPKw3bRtuYLCunX4dC+CkP2OqkfIMbuceGjUsJTMeWMISLYKSwuSpT2nmrzOnQ1ZOVGlt9dUbTPS/i1O9WAAlOmvMorpgzfYXYiaefwcI6cVJ05rGWy0Y5uPJWOI16lOzf5VAoGBAKpkBqqvPL7VXZWwTp/rvVmLcwzvApQydiHw12W2/IumvGO4UCttRshMVclNQklQPMLqVFL2UtB6H51dubC9roc0VnlwgTh1EFy/NGEs83PGLKOMU3vL7IBZeLX0ZY04ltjw5IoXiXoFv1ZKZmLr4E7ynTbi5x/NVIZ9JS+T0DWjAoGAGSKz3FRuSnzvZfohOQO3pG9S20ladWswTR2txUoxcyG5g5TJZSqHu/wal2DE41oHn0Y5TMqGyTzfGwwk+PC6LLT+PJGpXLpukX3Qg3JxIUy56MjRv2dhB/C0ErKxe28SXB+guGjcmdpx3acetE1qApzL1AewR47l3MNpF1u1D4k=",
+    "y": "f9ead70462aa9b4b656e83d3b7d2a8a13f6f2da09aa1ec2689da5a740f364cdd",
+    "s": {
+      "hash": "yUPfhNo7b5+FG0kgA2t93PtShV5qLaCkre7hqg6xEaCoLc3PEsG9OSS271Tf/vleOSnxz9rz01Zj/iPNjqY38JlTaB+OLfDdf7sPJ1hlBoRAoUuF+4WfCmTzuY8",
+      "s_key": "Secret Interpretation"
+    },
+    "z1": "aleo1defvj7u5yljwnyj4a6sluupxnk7gw34tzpjwd774zdaznlq2syxsphweak.private,",
+    "z2": "aleo1cnl4x5ldhyj6ru8uc8w2zkpjhl48gr9mgjdatu26x0ef5983w5yqtwedhe.private,",
+    "z3": "aleo1grhr6vvs02grvcxjzsqpasqysk2h0g7ked7e9yyaf26r0x0j7y8q9fd4dn.private,",
+    "z4": "aleo1rd6yte6yzjkymkurmy94xkyvujtrdnqcufpcezex7q7l335uquyqqz3rd6.private,",
+    "z5": "aleo1l4y00suwcu4r7y5q43r9rwfmkfvuvl036pz0vzhztc3jkpqaky8shdm3sq.private,",
+    "x1": {
+      "interpretation": "uS5lzt/AprdW6dOeYa4BVJWnOBfeltf4yVBQPXGQO69zQmpQX8GZY7/QyGSOfiaCvFHa3YhebZpFRyVuHOyYN3u6U5ueDrR8CmTg3+pXyMRuQYDZNw6ct5CNcBAQ9bdu5s1pRg",
+      "function_id": "purchase"
+    },
+    "x2": {
+      "interpretation": "Q79XlmYyYnhu7KTVwM+BcFR/e4DFaUdhf6nB30ZxLpYfgQdoCx3RQJRk59sZxZ043YkoPTjn8s85NMmDITpGjX6Wzm3BUESZjG7sD+hcCsSz75XJQT1umg+tC3pOuay7S3Pxnw",
+      "function_id": "recover"
+    },
+    "x3": {
+      "interpretation": "lcoIODrXmDUaGBNE928qjR87b9CbXnD+kvyFvBbV0BKwknGFv8WeJxpLKCLkKF2gFC7sH5xd4MSmD9Q9ov+1OJwsCogoR8rlgfiXeHa9pq6A+6HniGq0Nc5Obf43q45+qWbk/Q",
+      "function_id": "spend"
+    },
+    "x4": {
+      "interpretation": "Pc6rjVloS7JI3Gq5CpifFAopcXOMb2cTErIj8L+mshFqUlRDNvZmht5xth3f3hcH+A38nmfrVI84O8TK39aWrLBrxIRuyNIKoEbjmbiJ1R+/Pz/1RbVTy2b6X6M7ZzLNXjQMAQ",
+      "function_id": "transfer"
+    },
+    "x5": {
+      "interpretation": "eA6tjebiHHP776PLPc8ybXDCY/LLeE+zsIo32EeVaba5D4QoYlz7KW33qjsbwS8jnrLQTOkQ6zYW3LFWx7fUqBVVQhnyzxqblMq3GuiTnO4hwHNawM6iFvFLoZjOU1m+uFMisQ",
+      "function_id": "send"
+    }
+  },
+  "account_query": {
+    "node_id": {
+      "aleo_value": "aleo1tg2yvnn8r57lhjvqjkp6vl4206ddkfp0f0yx964hx8wlamzdf5ps063u9z.private,",
+      "tableset_value": "NODE ID"
+    },
+    "game_id": {
+      "aleo_value": "aleo1v0yrt50m4qkjjdg2anmls96sd8stkpnemap9kerev4tn2pu9zgps66eawj.private,",
+      "tableset_value": "GAME ID"
+    },
+    "default_currency": {
+      "query_value": "InvoBro",
+      "tableset_value": "DEFAULT CURRENCY"
+    },
+    "load_balance": {
+      "query_value": "192810",
+      "tableset_value": "LOAD BALANCE"
+    },
+    "pool_id": {
+      "aleo_value": "aleo1f045v8k6jgdfxgjgnaq86p08wlyhcrt4g7z68rle44305ykvdyxqhe7j0f.private,",
+      "tableset_value": "POOL ID"
+    },
+    "asset_id": {
+      "aleo_value": "6880614301973719965599342770527969173353052565868341678043217323101443991546field.private,",
+      "tableset_value": "ASSET ID"
+    },
+    "account_id": {
+      "aleo_value": "aleo1n685agqlmrfz6lzajca30jv9uqdc8qfr2sg7xyxz206gj4klyc8q2ag86g.private,",
+      "tableset_value": "ACCOUNT ID"
+    },
+    "gamertag": {
+      "query_value": "Test2",
+      "tableset_value": "GAMERTAG"
+    },
+    "public_key": null,
+    "txid": {
+      "aleo_value": "at1gmrd4p6dnrl325u70cqaeq6usvyzcla9yfzjtfy9cv84aqh07grs9j8hyk",
+      "tableset_value": "TRANSACTION ID"
+    },
+    "recordcipher": {
+      "aleo_value": "record1qyqsq2aq0y4cfzck8u6d6ah4k4n2jj6hyjs69kjy5ns5w8kq3s0a2pqsp5rkzurfta4k272rqqpqyq9y3cqgjlf67zg5hwqk43qx0e8kp8rp3vz7w7044l7793j7mmh3q4zuxtxvp6dtm95d3qh65qhtp20cq8lxlmkj6kp47uy0l0d6u32qgpmwdajx2hmfv3psqqszqqz29wmxgfngjqhhtsgeuh85gmvj7d9rj0t9ge28w7ppm8da57hsly565dn84swnzgv3zw4g6g8f7grt3qqvlep7d3yvtasucj4m7xgxqankzmt9ta5kgscqqgpqpacyh0mw9t8jtn9f22mcunk8eeuy8uzspm05qvf8496fevkyu3cjymcf72q4z2mkjdnp2kvxtfp7wdf042jumaqlh7gx9p9hgpnafgrqwur0dak976tygvqqyqsqg6mpkgcahcukdsqu4etdsf0xp96zj8qert353p0lt7a63tvff5ge5gyzh98a92lhd65v4mr28r27dvmymr2qak2425fh582uylt9urq2v93kxmm4de6976tygvqqyqsq7ckwmgy2fyky6p6cuxhuefdqvr3k4tpnhks6mycv4fc05w9vvsqf0apetszyedjt8um8pf4sawasdu2pvsff64r6lpap7qcl0ts06rggv9ehxet5ta5kgscqqgpqppm8n32c97kzpgn9wvkf03s5m8hmac8m0526a58wspj3m5j30jsryykjycg50rh5rkjjz0td5leexmwf6wuw8apm05m375unu9vnes9qkerfvenxje2lddjhjv2rqqpqyqqepp7d4s85m2u7k2wnl8lydh7l7q5g94w5kyhw5gzqqn03aghhpc2vyrmkffjpaw3stftv4tzgejw6fetmc2gjlsgpf49gj8nm6sg3yzmyd9nxv6t9ta4k27fjgvqqyqsqw2rl672e3u8mqzfm68xe0tpjpnpgxplg92jzcxjwtldas8429uqdcr69fyjy3g3vhsr0eppe9zxpugsxjrx7mh47twg2cme5h556wzsz0gc5xqqzqgqwae0g2szqswpf5ryq9ns0zjlqjurqjgfd5fa732kf30lfnudw7rr52tqtsa9yls0pklarknjjdh3pds3u4j9xp5n025cjccdez3p8qup85vjrqqpqyq8qwf4lm4npa9r6tkhdd435ux0nau8rtu35hfmmhfd2m7jcpq3ezt5vnejj6z9jljes6jlszch6hzsckdf44l0ld2mz9554zyf2md3syqn6xdpsqqszqpf6wjf0k0sdqgglj2h8k5cv9cthqr2ja73nu0ygn5seumwhdggqcxswv4awd4gjr9nrfqwkuqu7566aw43x3vl7qx5hnd8gyjfr3dq2qfargscqqgpqpyy0kr6ltgau55xnxnzxje52lfv6hxf8kfdz0fu6mx2a9g9pqeqphlxj28pjv8m43za72nwef9smsavrmsdygny0p4hxra3jfd8tw5zqy734gvqqyqsqyut5ejvsxdv6msfte24nykv3zlup5q7x8xuvragj6ju2xed7ssrvlx39tpuzrz5jhf7tdksnrv03r2ga7p4rrdx6stftg5hvqycysqh2wt6w8uzl6d80wpt9e8upe09nzwtxdzv09nr2jajh8h7vm92zpuu6rgp5",
+      "tableset_value": "RECORD CIPHER"
+    }
+  }
+}
