+++ src/stored_procedure/mod.rs
@@ -1,3 +1,5 @@
+pub mod errorzk;
+pub mod kellmen;
 pub mod keys;
 pub mod record;
